feat<AlbumList>专辑详情页与跳转逻辑

+++ core/data/src/main/java/com/example/data/apiService/AlbumList/AlbumListApiService.kt
@@ -7,5 +7,5 @@ import retrofit2.http.Query
 
 interface AlbumListApiService {
     @GET("album")
-    suspend fun getAlbumListData(@Query("id")id:Int): Response<AlbumListData>
+    suspend fun getAlbumListData(@Query("id")id: Long): Response<AlbumListData>
 }
