+++ core/data/src/main/java/com/example/data/repository/AlbumList/AlbumListRepository.kt
@@ -3,5 +3,5 @@ package com.example.data.repository.AlbumList
 import com.example.data.model.AlbumListData
 
 interface AlbumListRepository {
-    suspend fun getAlbumListData(id: Int): AlbumListData
+    suspend fun getAlbumListData(id: Long): AlbumListData
 }
